Parser is now a function assembling a dict.
The preliminary code was turned into a function that can now parse v1
deconvolution job files, assembles a dict structure with the collected
information and does some sanity checking.

Additionally, the job description file is adjusted to the new format
and a broken job file is supplied for testing. References #110