Remove "priority" field from job config files.
As stated in #265!#note-2 the priority should not be in the job config
itself, but rather derived by the Queue Manager from the knowledge it has
about the job and the system (which is much more detailed than the one
the PHP frontend has).

Plus, parsing of the priority field was never implemented in the QM part
anyway, for the same reasons as stated above.

refs #265, 136