Jobfile v7: introduce "tasktype" field in the "hucore" section.
This is to tell what kind of task hucore is going to process, and can be
used to derive in which queue a job should be placed, or which special
parameters should be supplied to the hucore-call.

Increases the job config file version string to "7", skipping the never fully
implemented version "6".

refs #265, #136, #269