Watch out for stuck jobs and remove them from the queue if necessary
This addresses GH-1 which is caused by the fact that GC3Pie is simply logging an
error message (`UnrecoverableDataStagingError`) in case input files are missing,
but doesn't pass on this information properly.

The idea of this workaround is to keep an eye on the number of jobs in status
`NEW` for a few cycles. In case they don't progress to status `RUNNING` we can
assume they are stuck and should be removed from the queue.

This also introduces a new job status `submitting`, indicating that a job has
been scheduled for execution by the spooler and is waiting to be picked up by
the gc3 engine.